Map the spread of SLF in the U.S.

The function takes a dataset of SLF occurrences and displays a map of the spread over time. The "zoom" option allows for a customized focus of the map

map_spread(
  resolution = "10k",
  zoom = "range",
  xlim_coord = NULL,
  ylim_coord = NULL,
  rarefy = TRUE,
  color_palette = "plasma",
  print_message = TRUE
)

Arguments

resolution

Defines the resolution at which the data is displayed. `"1k"` resolution shows data aggregated at a 1km2 grid, while `"10k"` shows tiles aggregated at a 10km2 grid

zoom

Defines the limits of the map to be plotted. `"range"` limits the map to the established range for SLF; `"full"` displays the whole United States. `"custom"` the user to specify the range over which the map should be displayed (see `xlim_coord, ylim_coord` below).

xlim_coord, ylim_coord

Ordered numerical vectors of length 2 determining the longitudinal (`xlim`) and latitudinal (`ylim`) boundaries of the map, to be expressed as decimal degree coordinates. Unnecessary if zoom is set to any value other than `"custom"`

rarefy

Logical value (`TRUE/FALSE`). Specifies whether the data should be summarized for faster plotting. Overwritten and set to `TRUE` when zoom is full and/or the resolution is 10k.

color_palette

A character string indicating the color option to use. Five options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E").

print_message

Logical value. Can be set to `FALSE` to shut down the warning on run time.

Value

A ggplot of the spread of SLF

Examples

## Examples